% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_result_data.R
\name{write_gsea_result_json}
\alias{write_gsea_result_json}
\title{Write Shared GSEA Data to JSON}
\usage{
write_gsea_result_json(gsea_result, path)
}
\arguments{
\item{gsea_result}{Data returned by \code{\link[=gsea_result_data]{gsea_result_data()}}.}

\item{path}{Output JSON file.}
}
\value{
\code{path}, invisibly.
}
\description{
Write Shared GSEA Data to JSON
}
